Split up Tinyxml2Serializer/Tinyxml2Deserializer
