+++ CMakeLists.txt
@@ -16,7 +16,8 @@ add_library(${PROJECT_NAME}
         Settings.cpp
         MarkupSerializer.cpp
         Tinyxml2Serializer.cpp
-)
+        Tinyxml2Deserializer.cpp
+ )
 
 target_include_directories(${PROJECT_NAME} PUBLIC
         $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>
