+++ RTProtocol.h
@@ -10,11 +10,9 @@
 #include <string>
 #include <map>
 #include <limits>
-#include <cmath>
 #include <cstdint>
 #include <memory>
 
-#include "MarkupSerializer.h"
 #ifdef _WIN32
 #pragma warning (disable : 4251)
 #endif
